feat: Rework of configuration structure
As the current config structure was a bit complicated to use, it has
been completely reworked.

First of all, there is now a central manifest file, which specifies all
the details needed for deployment. This includes a list of all available
projects (including paths to them, so goodbye recursive project
discovery) and a grouping of available environment (including the name
of the env variable holding the token).

The structure within a project now has to follow a strict scheme,
naminly `[api]/[config]`. Auto-discovery of sub-Projects are no
longer supported. This was done to make things more explicit.

The configuration yaml also looks now different. Instead of splitting
configuration definitions in two parts (template location, rest of the
config), everything is now in one structure. The top level structure of
each yaml file is now called `configs`. For examples please see the
documentation.

Config paramters also have been reworked. They now support more values
besides strings. There are some special parameter though. One of this
special parameter is the reference parameter. References no longer
depend on a special formatted string, instead you have to explicitly
define the location of the config you want to reference. It is now also
possible to depend on **any** reference of a config, not just id and
name.

Another special parameter is the environment parameter. It is **no longer**
possible to reference env variables via the `{ .env.XYZ }` syntax in the
config yaml or the template. This decision was made, to make configs
more explicit.

For more details on how the new parameter syntax works, please see the
documentation.

The delete.yaml now also supports deleting configs with multiple `/` in
its name.